python/requirements.txt
python-dotenv
praw
minio
google-api-python-client
